first working update.sh

+++ update.sh
@@ -1,35 +1,38 @@
 #!/bin/bash
-#
-# Script name: build-packages.sh
-# Description: Script for automating 'makepkg' on the PKGBUILDS.
+
+# Source #
 # GitLab: https://www.gitlab.com/dwt1/dtos-pkgbuild
 # Contributors: Derek Taylor
 
-# Set with the flags "-e", "-u","-o pipefail" cause the script to fail
-# if certain things happen, which is a good thing.  Otherwise, we can
-# get hidden bugs that are hard to discover.
 set -euo pipefail
 echo "$0"
 
 
 REPO_NAME="arch-repo"
 REPO_URL="https://github.com/FelixFeuerigel/arch-repo.git"
+REPO_DIR="../$REPO_NAME"
+CHROOT="$PWD/chroot-root"
 
-CHROOT="$PWD/root"
+
+PKG_DIR="$(pwd)"
 
 echo "Enter a commit mesage: "
 read COMMIT_MESSAGE
 
-
-if [ -d "../$REPO_NAME" ]; then
-    cd ..
+if [ ! -d "$REPO_DIR" ]; then
     echo "###########################"
     echo "Cloaning the repo database."
     echo "###########################"
-    git clone $REPO_URL
-    cd -
-fi
+    git clone $REPO_URL $REPO_DIR
+else
+    echo "###########################"
+    echo "Updating the repo database."
+    echo "###########################"
 
+    cd $REPO_DIR
+    git pull || echo "==> WANING: Failed to pull from git remote."
+    cd $PKG_DIR
+fi
 
 ## make chroot evironment if needed
 mkdir -p "$CHROOT"
@@ -37,37 +40,37 @@ mkdir -p "$CHROOT"
 
 ## build packages in "x86_64" directory
 readarray -t x86_list <<< "$(find x86_64/ -type f -name PKGBUILD | awk -F / '{print $2}')"
-echo "######################"
-echo "Building ${#array[@]} packages."
-echo "######################"
+echo "#####################"
+echo "Building ${#x86_list[@]} packages."
+echo "#####################"
 for x in "${x86_list[@]}"; do
     cd x86_64/"${x}"
-    echo "### Making Package: ${x} ###"
-    updpkgsums PKGBUILD || echo "FAILED TO UPDATE CHECKSUMS OF PACKAGE: ${x}"
-    makechrootpkg -cur $CHROOT -- -cf || echo "FAILED TO MAKE PACKAGE: ${x}"
-    # makepkg -cf --sign || echo "FAILED TO MAKE PACKAGE: ${x}"
+    echo -e "\n ### Making Package: ${x} ###"
+    updpkgsums PKGBUILD || ( echo "ERROR: FAILED TO UPDATE CHECKSUMS OF PACKAGE: ${x}" && exit 1 )
+    makechrootpkg -cur $CHROOT -- -cf || ( echo "ERROR: FAILED TO MAKE PACKAGE: ${x}" && exit 1 )
+    # makepkg -cf --sign || echo "FAILED TO MAKE PACKAGE: ${x}"  && exit 1
     # find . -mindepth 1 -maxdepth 1 -type d -print0 | xargs -r0 rm -R
-    cd -
+    cd $PKG_DIR
 done
 
-x86_pkgbuild=$(find /x86_64 -type f -name "*.pkg.tar.zst*")
+echo "################################################"
+echo "Moving the packages to the \"$REPO_NAME\" repo."
+echo "################################################"
+
+[[ -d "$REPO_DIR/x86_64/" ]] || mkdir $REPO_DIR/x86_64/
+
+x86_pkgbuild=$(find x86_64/ -type f -name "*.pkg.tar.zst*")
 
 for x in ${x86_pkgbuild}; do
-    echo "Moving ${x} to $REPO_NAME/x86_64"
-    mv "${x}" ../$REPO_NAME/x86_64
+    echo "Moving ${x} to $REPO_DIR/x86_64/"
+    mv "${x}" $REPO_DIR/x86_64/
 done
 
 git add .
-#git commit -m "$COMMIT_MESSAGE"
-#git push
+git commit -m "$COMMIT_MESSAGE"
+git push
 
-cd ../$REPO_NAME/x86_64
-
-echo "###########################"
-echo "Updating the repo database."
-echo "###########################"
-
-git pull
+cd $REPO_DIR/x86_64
 
 echo "###########################"
 echo "Building the repo database."
@@ -85,17 +88,17 @@ echo "############################################"
 ## -R: remove old package files when updating their entry
 repo-add -s -n -R $REPO_NAME.db.tar.gz *.pkg.tar.zst
 
-# Removing the symlinks because GitLab can't handle them.
+## Removing the symlinks because GitLab can't handle them.
 rm $REPO_NAME.db
-rm $REPO_NAME.db.sig
+[[ -d $REPO_NAME.db.sig ]] && rm $REPO_NAME.db.sig
 rm $REPO_NAME.files
-rm $REPO_NAME.files.sig
+[[ -d $REPO_NAME.files.sig ]] && rm $REPO_NAME.files.sig
 
-# Renaming the tar.gz files without the extension.
+## Renaming the tar.gz files without the extension.
 mv $REPO_NAME.db.tar.gz $REPO_NAME.db
-mv $REPO_NAME.db.tar.gz.sig $REPO_NAME-db.sig
+[[ -d $REPO_NAME.db.tar.gz.sig ]] && mv $REPO_NAME.db.tar.gz.sig $REPO_NAME-db.sig
 mv $REPO_NAME.files.tar.gz $REPO_NAME.files
-mv $REPO_NAME.files.tar.gz.sig $REPO_NAME.files.sig
+[[ -d $REPO_NAME.files.tar.gz.sig ]] && mv $REPO_NAME.files.tar.gz.sig $REPO_NAME.files.sig
 
 echo "################################"
 echo "Uploading the database git repo!"
@@ -109,4 +112,4 @@ echo "#######################################"
 echo "Packages in the repo have been updated!"
 echo "#######################################"
 
-## TODO: After testing package functionality check it for errors using namcap
+cd $PKG_DIR
